Voiceprint Analysis/requirements.txt
# Core Deep Learning Frameworks
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# Speaker Recognition Models
speechbrain==0.5.16
transformers==4.35.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10
noisereduce==3.0.0

# Feature Extraction
numpy==1.24.3
scipy==1.11.4

# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Security and Encryption
cryptography==41.0.7
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
pymongo==4.6.0
motor==3.3.2

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Monitoring and Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

